pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "betulon"
dependencies = [
    "mastodon.py==1.8.1"
]
requires-python = ">=3.10"
authors = [
    {name = "Giovanni Lanzani", email = "pypi@lanzani.nl"},
]
maintainers = [ 
    {name = "Giovanni Lanzani", email = "pypi@lanzani.nl"},
]
description = "Project to import Mastodon bookmarks to Betula"
dynamic = ["version"]

[project.scripts]
betulon = "betulon.main:cli"

[project.optional-dependencies]
dev = ["pytest>=7.4.3"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"

